Change ret_code from __u8 to __s64 for flexibility
Per review feedback, change struct landlock_supervise_response to use
__s64 for ret_code (matching seccomp-unotify) and __u16 for flags.
New layout:
- __u16 length
- __u16 flags
- __u32 cookie
- __s64 ret_code (total size: 16 bytes)

Updated kernel event structure, supervisor.c, supervisor_sandboxer.c,
and design.md to reflect the new field types and layout.

+++ security/landlock/supervisor.c
@@ -346,12 +346,10 @@ static void landlock_supervise_wait_work(struct callback_head *twork)
 	/* Set the syscall return value based on the supervisor's response. */
 	if (event->response_flags & LANDLOCK_SUPERVISE_RETCODE) {
 		/*
-		 * Supervisor specified a return code. The ret_code field is u8,
-		 * but represents a signed error code. Sign-extend from s8 to long
-		 * to properly handle negative error codes like -EPERM (stored as 0xF1).
+		 * Supervisor specified a return code directly.
 		 */
 		syscall_set_return_value(current, regs,
-					 (long)(s8)event->response_ret_code,
+					 (long)event->response_ret_code,
 					 0);
 	} else {
 		/* Default behavior: restart the syscall. */
